Initial Model-View-Controller refactor commit.
After weeks of isolation, the refactored code enters the repository. Expect a
lot of changes and buggy code. The code is based as much as possible on the
trunk code to prevent extra work.

Main reasons to rewrite the GUI have been that the original defo GUI is fixed
size and everything was placed in one file: DefoMainWindow.cc. The code is now
being split up in smaller, logical units.

Different models that were written to interface with the devices and signal the
GUI elements connected to the model of any changes in the state of the device.

List of models:
    MODEL             DEVICE
  * DefoCameraModel   Canon EOS 550D DSLR
  * DefoConradModel   Conrad Relais Switch
  * DefoJulaboModel   Julabo FP50 chiller
  * DefoKeithleyModel Keithley 2700 multimeter (temperature read out)

Each of the models is accompanied by one or more test widgets. Testing is
performed by adding them in TestWindow

Up to now the main refactoring focus has been on the device interfacing and
removing unneeded code from trunk.