Preparation for visual inspection of camera setup configuration.
DefoMeasurement is adjusted to only contain immutable information obtained at measurement. Any offline processing will be stored in seperate classes (still to be done).

DefoSurfaceModel was adjusted to contain the list of measurements and a single selection. A reference and measurement for deformation calculation must thus be stored in a seperate model (not yet available).

DefoImageWidget was abstracted and subclassed twice to provide widgets that display the raw camera image (DefoRawImageWidget) and the threshold filters (DefoImageThresholdWidget).

Removed some redundant files: DefoImageLabel.*, defo.ui

TestWindow currently does not compile, the reason for this is not entirely clear.