Measurement selection and multi threaded point recognition. The current program allows for picture taking and dot finding.
The MeasurementListModel now only stores measurements, selection is now handled by MeasurementSelectionModel. This allows for easy selection of different measurements for different views and controllers; as is needed for comparison to a reference measurement.
This new model is accompanied by a new (test) view MeasurementListComboBox.

The MeasurementListModel now also stores the cache for retrieved points and is made thread safe to allow for parallel point recognition on different image parts. See also PointFinder in TestWindow.*

The class CoordinateSaver in TestWindow.* writes out the found points of the current selection to a text file for basic processing of their properties. As such it was found that, given a good alignment of the dot grid with the image axises, point binning is feasible. This was already used in the original code (see DefoRecoSurface).