Create driver files for Velleman VM 8090 relay card on Linux
The DESY CMS Group that created the PS automated assembly process
used a relay card model made by Conrad which was not available
in the United States. As such, the Brown CMS Group chose the
Velleman VM 8090 relay card as a replacement model in their
set-up. This code was created to allow the VM 8090 relay card
to interface with the computer on a Linux OS. It was modeled on
the code for the Conrad model relay card, which can be found at
https://github.com/DESY-FH-ELab/cmstkmodlab/tree/master/devices/
Conrad.