Improve ability to test and effectiveness of VM 8090 Linux driver files
Files were modified to facilitate their successful compilation. Changes
were also made to improve functionality and readability of error
messages based on testing the driver capabilities with the /utils
/vellemanTest script, the compilation instructions of which were added
to Makefile.in. The setMultiRelays method in the IVellemanController
interface and all derived classes was amended to allow for setting
multiple relays off (or on, still) while not affecting any other relays
(as opposed to turning multiple relays on and all other relays off) in
the VellemanController class. .gitignore files were added to the main
Velleman folder and the /utils folder.