Fix compilation issue stemming from a call to LStepExpressFake::ioPort()
The ioPort() method was not previously implemented in the LStepExpressFake
module, though it was/is implemented in the LStepExpress module. As a
result, calls to ioPort() from an object that could either be an instance
of LStepExpress OR LStepExpressFake raised compilation errors. An ioPort()
method was added to the LStepExpressFake module to resolve this problem.