Standardize references to Velleman VM 8090 relay card
The VellemanController, VellemanModel, and VellemanManager modules
were inconsistent in their indexing of relays and use of the
word 'relay' in methods. I standardized the indexing/references to
relay or channel numbers in methods so that the first channel is
ALWAYS called with '1,' the second with '2,' etc. Furthermore, I
changed the names of methods in VellemanModel and VellemanManager
to refer to 'channel(s)' instead of 'relay(s)' in order to
stay consistent with vacuum tools independent of the relay card
interfacing files in the software.