Improve assembly program's communication with Velleman VM 8090 relay card
A descriptive error message was added to the intialization method of
the VellemanModel class as defined in /common/VellemanModel.cc. In addition,
the readStatus private VellemanController class method was modified such
that its arguments are now references to unsigned characters instead of
unsigned characters. This change repairs a memory leak in reporting the
statuses of relays in the Velleman VM 8090 relay card. Finally, commented
questions were removed from  the end of the VellemanManager class header
file in the /assembly/assemblyCommon directory as the DESY group said the
questions had little current relevance.