Add fourth vacuum line capability in main assembly script
The AssemblyVacuumWidget, AssemblyAssembly, and AssemblyHardwareControlView
modules in the /assembly/assemblyCommon directory were modified to
allow for the use of a fourth vacuum line to the secondary stage in
addition to the baseplate, spacers, and pickup tool vacuum lines.
The main assembly configuration file was also changed to reflect
this new functionality, and a new assembly procedure configuration
script was added to reflect the current Brown PS automation set-up.