Adjust settings after axis orientation standardization
The orientation of the Z- and A-axes of the motion and rotation
stages were found to be inconsistent between Brown and DESY. Brown
adapted its orientation settings to align with those of DESY
(positive Z is upwards towards the ceiling, while position A is
counterclockwise). Due to these changes, assembly configuration
parameters and references points were updated as well. Furthermore,
work-arounds placed in certain assembly support files to deal with
the previous differences in axes orientations were removed.