Enhance and update alignment and assembly capabilities
Due to changes in the surroundings of the PS automation set-up,
including the addition of a cable tray to organize wires and tubing
around the set-up, changes to assembly settings and parameters were
also necessary. Additional modifications were made in attempts to
reduce variation in angular and linear misalignment between sensors
in dummy modules assembled using PS automation.